update to error on compile error

+++ java_runner.py
@@ -77,7 +77,8 @@ class JavaRunner(CodeRunner):
     def cleanup(self, source: str):
         os.remove(source)
         classname = os.path.splitext(source)[0] + ".class"
-        os.remove(classname)
+        if os.path.exists(classname):
+            os.remove(classname)
         os.rmdir(source[:source.rfind("/")])
